test(e2e): unit-test hydration quiescence helper outside ci-smoke
Move the #7212 false-green/false-red quiescence contracts into a Node
unit test and shared helper so variant-smoke-full no longer pays two
extra Playwright boots for them.
